requirements.txt
# Web Framework & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
numba==0.58.1

# Machine Learning & Math
scikit-learn==1.3.2
numpy==1.26.3
pandas==2.2.0

# File Handling
joblib==1.3.2